// src/service/file/schemas/cdnFileDownloadRequest.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

export interface CdnFileDownloadRequest {
  /** 다운로드를 하고자하는 자료들의 키 값 리스트 */
  driveItemKeys: string[];
  /** url의 유지할 분 수 */
  period: number;
  /**
   * 요청자의 ip
   * @minLength 1
   * @maxLength 40
   */
  ip: string;
  /**
   * 다운로드용 링크 요청읹지? 미리보기라면 false
   * @nullable
   */
  isForDownload: boolean | null;
  forDownload?: boolean;
}